Refactor Code to use functions
